REFACTOR: I took out the extractCertificateFromReq function from the certificate controller to enhance code readabilty
